requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
snowflake-connector-python==3.5.0
boto3==1.34.0
pydantic==2.5.0
pydantic-settings==2.1.0
mangum==0.17.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
python-multipart==0.0.6
